requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.5
idna==3.4
iniconfig==2.0.0
numpy==1.23.3
packaging==23.1
pandas==1.5.0
parseargs==0.0.5
Pillow==9.2.0
pluggy==1.0.0
pytest==7.3.1
python-dateutil==2.8.2
pytz==2022.4
qrcode==7.3.1
regex==2023.5.5
six==1.16.0
urllib3==1.26.12